remove ability to use karma-requirejs
There are quite some problems using this plugin with karma-requirejs.
Using this without require works much more reliable.

However, you have to be careful about the order of the frameworks you
register with karma. Make sure to register 'chai' after
'chai-as-promised', because chai-as-promised depends on chai to be
loaded. The next step will be to handle the load order in a more
intelligent way.